FlyRL.preprocess — Method

preprocess(
    df;
    filter,
    window,
    initial_outlier_tolerance,
    encoders,
    warn_outliers,
    drop_outliers,
    subsample_resolution
)

Preprocess data frame df.

FlyRL.read — Method

read(root, f; preprocess, kwargs...)

Read track f in directory root and corresponding time, shock and pattern file. Returns a DataFrame. kwargs are passed to preprocess.

FlyRL.read_directory — Method

read_directory(dir; verbosity, pattern, kwargs...)

Read all tracks in directory dir. Uses read and passes kwargs to preprocess.

FlyRL.AngleEncoder — Type

AngleEncoder()

Encodes angles between subsequent orientations.

Convention: the same row should contain the delta values that lead to the current value, i.e. delta values should start with missing.

FlyRL.AngleEncoder2D — Type

AngleEncoder2D()

Encodes angles as (sin(angle), cos(angle)) (see also AngleEncoder).

FlyRL.ArmEncoder — Method

ArmEncoder(; with_outliers)

Encodes arms of the maze as "left", "middle", "right", "center". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.ClosestWallEncoder — Type

ClosestWallEncoder(encoder)

Encodes the closest wall as determined with encoder (either FourWallsEncoder, SixWallsEncoder or EightWallsEncoder).

FlyRL.ColorEncoder — Method

ColorEncoder(; colordict, with_outliers)

Encode color based on track.pattern.

FlyRL.ColumnPicker — Type

ColumnPicker(colname::Symbol)

Picks a column from a data frame.

FlyRL.DeltaPositionEncoder — Type

DeltaPositionEncoder()

Encodes Δx and Δy from x and y.

Convention: the same row should contain the delta values that lead to the current value, i.e. delta values should start with missing.

FlyRL.DeltaPositionIndexEncoder — Type

DeltaPositionIndexEncoder(; maxradius = 20, stepsize = 1)

Encodes relative positions (Δx, Δy) with an integer, based on a 2D grid with stepsize and largest relative position Δx^2 + Δy^2 = maxradius^2.

FlyRL.DeltaTimeEncoder — Type

DeltaTimeEncoder()

Encodes Δt from t; uses default 0.15, if t is missing.

Convention: the same row should contain the delta values that lead to the current value, i.e. delta values should start with missing.

FlyRL.DurationPerStateEncoder — Type

DurationPerStateEncoder(encoder)

Encode the duration that is spent in a given state.

Example

encode(DurationPerStateEncoder(ShockArmEncoder()), random_track()) |> DataFrame

FlyRL.DynamicCompressEncoder — Method

DynamicCompressEncoder(compress_on, encoders...)

Merge subsequent states of compress_on into one state.

Example

encode(DynamicCompressEncoder(StanHMCAdaptor()), random_track()) |> DataFrame

FlyRL.FutureDeltaTimeEncoder — Type

FutureDeltaTimeEncoder()

Same as DeltaTimeEncoder but offset by one, i.e. the same row contains the delta value that leads to the next value.

FlyRL.InShockArmEncoder — Type

InShockArmEncoder()

Encodes if the position is in the shock arm.

FlyRL.LevelEncoder — Type

LevelEncoder(encoder)

Encode level as an integer for a categorical encoder (like ShockArmEncoder).

FlyRL.MarkovKEncoder — Method

MarkovKEncoder(K, encoders)

Create copies of the original encoders with offsets up to K. Appends ˌk to copy with offset k.

Example

encode(MarkovKEncoder(3, DynamicCompressEncoder(:arm, ArmEncoder())), random_track()) |> DataFrame

FlyRL.OrientationEncoder — Type

OrientationEncoder()

Encodes orientations based on Δx and Δy.

FlyRL.SemanticEncoder3 — Method

SemanticEncoder3(; with_outliers)

Encodes positions of the maze as "center", "arm" or "turn". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.SemanticEncoder7 — Method

SemanticEncoder7(; with_outliers)

Encodes positions of the maze as "center", "left arm", "left turn", "middle arm", "middle turn", "right arm", "right turn". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.ShockArmEncoder — Method

ShockArmEncoder(; with_outliers)

Encodes arms of the maze as "neutral A", "neutral B", "shock" and "center". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.SpeedEncoder — Method

SpeedEncoder(; outlier_threshold, return_outliers)

Encode the speed based on vx and vy (see VelocityEncoder). Speed values above outlier_threshold are returned as missing. If return_outliers = true, return also a vector of outlier values.

FlyRL.VectorEncoder — Method

VectorEncoder(encoders; T, dropmissing, intercept)

Encode multiple encodings in a single vector. Categorical encoders are one-hot encoded.

Example

encode(VectorEncoder(ShockArmEncoder(), SpeedEncoder()), random_track(N = 10)) |> DataFrame

FlyRL.VelocityEncoder — Type

VelocityEncoder()

Encodes velocities based on Δx, Δy and Δt.

FlyRL.EightWallsEncoder — Method

EightWallsEncoder(; max_distance = 40, scale = max_distance, stepsize = 1.5)

Encodes (x, y, ox, oy)-position-orientation tuples to relative distances to the next wall at 0, 45, 90, 135, 180, 225, 270 and 315 degrees relative to the orientation.

If the distance of the wall is max_distance and scale == max_distance the returned value is 1. The stepsize (in pixels) controls the granularity at which the distance to the wall is computed.

FlyRL.FourWallsEncoder — Method

FourWallsEncoder(; max_distance = 40, scale = max_distance, stepsize = 1.5)

Encodes (x, y, ox, oy)-position-orientation tuples to relative distances to the next wall at 0, 90, 180 and 270 degrees relative to the orientation.

If the distance of the wall is max_distance and scale == max_distance the returned value is 1. The stepsize (in pixels) controls the granularity at which the distance to the wall is computed.

FlyRL.SixWallsEncoder — Method

SixWallsEncoder(; max_distance = 40, scale = max_distance, stepsize = 1.5)

Encodes (x, y, ox, oy)-position-orientation tuples to relative distances to the next wall at 0, 45, 135, 180, 225 and 315 degrees relative to the orientation.

If the distance of the wall is max_distance and scale == max_distance the returned value is 1. The stepsize (in pixels) controls the granularity at which the distance to the wall is computed.

FlyRL.decode — Method

decode(e, vectors)

Inverse of encode(VectorEncoder(), track).

FlyRL.encode! — Method

Encode track with encoder and store result in data frame track.

FlyRL.encode — Method

Encode track with encoder.

FlyRL.DenseLayer — Method

DenseLayer(Din, Dout, f; T, w, b)

Makes a dense layer.

FlyRL.MLP — Method

MLP(layers)

Makes a multi-layer perceptron.

Example

mlp = MLP(DenseLayer(10, 20, relu), DenseLayer(20, 1, identity))
mlp(rand(10))

FlyRL.train — Function

train(agent, data)
train(
    agent,
    data,
    params;
    epochs,
    verbosity,
    opt,
    print_interval,
    lopt,
    maxtime,
    optim_options,
    maxeval,
    lb,
    ub,
    fixed,
    batchsize
)

Train agent on preprocessed data with initial parameters params (=params(model) by default).

Keep the default opt = :MLSL for small scale problems and set opt = Adam(), batchsize = 32, if training on minibatches of size 32 is desired. Fix parameter values with e.g. fixed = (; η = 0.). Gradient evaluation is multi-threaded, if julia is started with multiple threads, e.g. bash> julia -t8.

FlyRL.train — Function

train(model, data)
train(model, data, params; kwargs...)

Train a model to a multiple data frames data with initial parameters params (=params(model) by default).

FlyRL.train — Function

train(model, data)
train(
    model,
    data,
    params;
    multi_threading,
    nthreads,
    kwargs...
)

Train a model to a single data frame data with initial parameters params (=params(model) by default).

FlyRL.DeltaPositionModel — Method

DeltaPositionModel(
;
    nhidden,
    σ,
    T,
    dx,
    dy,
    f,
    encoder,
    kwargs...
)

Make a DeltaPositionModel.

FlyRL.Model — Type

Model(agent, preprocessor)

Makes a model.

FlyRL.PolicyGradientAgent — Method

PolicyGradientAgent(; Din, Dout, T, model, update_lag)

FlyRL.StationaryAgent — Method

StationaryAgent(; Dout, model, T)

FlyRL.logprob — Method

logprob(model, data, params)

Compute the log-probability of the data given the model with parameters params.

FlyRL.Preprocessor — Method

Preprocessor(; input, target, shock, target_offset)

Define a preprocessor with input encoder input, target encoder target and shock encoder shock (= ColumnPicker(:shock) by default).

FlyRL.preprocess — Method

preprocess(p, data)

Preprocess data with preprocessor p (see also Preprocessor).

FlyRL.plot_compare_probs — Method

plot_compare_probs(track1, track2; ranges, f, kwargs...)

Compare the probabilities of track1 and track2 for being in certain states at different temporal ranges. The default ranges are (0, 1) (full range), (0, .5) (first half) (.5, 1)) (second half). kwargs are passed to plot_probs.

FlyRL.plot_maze — Method

plot_maze(track; f)

Plot the outline of a maze and the positions of a fly (track.x and track.y). See also plot_track. f = Makie.Figure().

FlyRL.plot_probs — Method

plot_probs(track; preprocessor, kwargs...)

Plot probabilities (frequencies) of being in certain states. By default the states defined by the ShockArmEncoder are taken, i.e. preprocessor = Preprocessor(input = DynamicCompressEncoder(:shock_arm, ShockArmEncoder()), target = ShockArmEncoder())

FlyRL.plot_summaries — Method

plot_summaries(tracks; f, summaries)

Plot some summary statistics for a collection of tracks.

FlyRL.plot_summary — Method

plot_summary(summary, tracks; f)

FlyRL.plot_track — Method

plot_track(track; f)

Plots a track as a function of time (see plot_track_3D) and on the maze outline (see plot_maze).

FlyRL.plot_track_3D — Method

plot_track_3D(
    track;
    f,
    cm_per_pixel,
    ylabel,
    xlabel,
    zlabel,
    perspectiveness,
    elevation,
    azimuth,
    viewmode,
    alignmode,
    protrusions,
    yticks,
    zticks
)

Plots track.x and track.y as a function of track.t.

FlyRL.ChangeOf — Type

ChangeOf(stat)
ChangeOf(stat, midpoint)

Compute the change in statistics stat before and after the midpoint (= 0.5 by default).

Example

summarize(ChangeOf(RelativeTimeInShockArm()), track)

FlyRL.RelativeTimeInState — Method

RelativeTimeInState(encoder, state; exclude)

Compute the relative duration spent in state in the encoding defined by the encoder. States can be excluded with exclude = ["state1", "state2"].

Example

summarize(RelativeTimeInState(SemanticEncoder7(), "left turn", exclude = ["center"]), track)

FlyRL.RelativeVisitsToState — Method

RelativeVisitsToState(e, key; exclude)

Compute the relative number of visits to state in the encoding defined by the encoder. States can be excluded with exclude = ["state1", "state2"].

Example

summarize(RelativeVisitsToState(SemanticEncoder7(), "left turn", exclude = ["center"]), track)

FlyRL.RelativeTimeInShockArm — Method

RelativeTimeInShockArm()

Relative time in shock arm.

FlyRL.RelativeVisitsToShockArm — Method

RelativeVisitsToShockArm()

Relative visits to shock arm.

FlyRL.Environment — Type

Environment(; preprocessor,
              pattern = "GBB",
              shock = shock_function(pattern),
              state = rand_state(preprocessor, pattern, shock))

Define an environment that can be used for simulations.

FlyRL.random_track — Method

random_track(; N, pattern, shock)