Encoders can be used to transform the raw data in different ways. This can be done in two ways, illustrated here with the ShockArmEncoder:

Create a new data frame with the transformed data

using DataFrames
import FlyRL: encode, ShockArmEncoder, random_track
track = random_track(N = 100)
encode(ShockArmEncoder(), track) |> DataFrame

Append the transformed data directly to the original data frame

using DataFrames
import FlyRL: encode!, ShockArmEncoder, random_track
track = random_track(N = 100)
encode!(ShockArmEncoder(), track)

Here is a list of built-in encoders.

FlyRL.AngleEncoder — Type

AngleEncoder()

Encodes angles between subsequent orientations.

Convention: the same row should contain the delta values that lead to the current value, i.e. delta values should start with missing.

FlyRL.AngleEncoder2D — Type

AngleEncoder2D()

Encodes angles as (sin(angle), cos(angle)) (see also AngleEncoder).

FlyRL.ArmEncoder — Method

ArmEncoder(; with_outliers)

Encodes arms of the maze as "left", "middle", "right", "center". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.ClosestWallEncoder — Type

ClosestWallEncoder(encoder)

Encodes the closest wall as determined with encoder (either FourWallsEncoder, SixWallsEncoder or EightWallsEncoder).

FlyRL.ColorEncoder — Method

ColorEncoder(; colordict, with_outliers)

Encode color based on track.pattern.

FlyRL.ColumnPicker — Type

ColumnPicker(colname::Symbol)

Picks a column from a data frame.

FlyRL.DeltaPositionEncoder — Type

DeltaPositionEncoder()

Encodes Δx and Δy from x and y.

Convention: the same row should contain the delta values that lead to the current value, i.e. delta values should start with missing.

FlyRL.DeltaPositionIndexEncoder — Type

DeltaPositionIndexEncoder(; maxradius = 20, stepsize = 1)

Encodes relative positions (Δx, Δy) with an integer, based on a 2D grid with stepsize and largest relative position Δx^2 + Δy^2 = maxradius^2.

FlyRL.DeltaTimeEncoder — Type

DeltaTimeEncoder()

Encodes Δt from t; uses default 0.15, if t is missing.

Convention: the same row should contain the delta values that lead to the current value, i.e. delta values should start with missing.

FlyRL.DurationPerStateEncoder — Type

DurationPerStateEncoder(encoder)

Encode the duration that is spent in a given state.

Example

encode(DurationPerStateEncoder(ShockArmEncoder()), random_track()) |> DataFrame

FlyRL.DynamicCompressEncoder — Method

DynamicCompressEncoder(compress_on, encoders...)

Merge subsequent states of compress_on into one state.

Example

encode(DynamicCompressEncoder(ShockArmEncoder()), random_track()) |> DataFrame

FlyRL.FutureDeltaTimeEncoder — Type

FutureDeltaTimeEncoder()

Same as DeltaTimeEncoder but offset by one, i.e. the same row contains the delta value that leads to the next value.

FlyRL.InShockArmEncoder — Type

InShockArmEncoder()

Encodes if the position is in the shock arm.

FlyRL.LevelEncoder — Type

LevelEncoder(encoder)

Encode level as an integer for a categorical encoder (like ShockArmEncoder).

FlyRL.MarkovKEncoder — Method

MarkovKEncoder(K, encoders)

Create copies of the original encoders with offsets up to K. Appends ˌk to copy with offset k.

Example

encode(MarkovKEncoder(3, DynamicCompressEncoder(:arm, ArmEncoder())), random_track()) |> DataFrame

FlyRL.OrientationEncoder — Type

OrientationEncoder()

Encodes orientations based on Δx and Δy.

FlyRL.SemanticEncoder3 — Method

SemanticEncoder3(; with_outliers)

Encodes positions of the maze as "center", "arm" or "turn". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.SemanticEncoder7 — Method

SemanticEncoder7(; with_outliers)

Encodes positions of the maze as "center", "left arm", "left turn", "middle arm", "middle turn", "right arm", "right turn". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.ShockArmEncoder — Method

ShockArmEncoder(; with_outliers)

Encodes arms of the maze as "neutral A", "neutral B", "shock" and "center". Points outside the maze are marked as "outlier" if with_outliers = true.

FlyRL.SpeedEncoder — Method

SpeedEncoder(; outlier_threshold, return_outliers)

Encode the speed based on vx and vy (see VelocityEncoder). Speed values above outlier_threshold are returned as missing. If return_outliers = true, return also a vector of outlier values.

FlyRL.VectorEncoder — Method

VectorEncoder(encoders; T, dropmissing, intercept)

Encode multiple encodings in a single vector. Categorical encoders are one-hot encoded.

Example

encode(VectorEncoder(ShockArmEncoder(), SpeedEncoder()), random_track(N = 10)) |> DataFrame

FlyRL.VelocityEncoder — Type

VelocityEncoder()

Encodes velocities based on Δx, Δy and Δt.

FlyRL.EightWallsEncoder — Method

EightWallsEncoder(; max_distance = 40, scale = max_distance, stepsize = 1.5)

Encodes (x, y, ox, oy)-position-orientation tuples to relative distances to the next wall at 0, 45, 90, 135, 180, 225, 270 and 315 degrees relative to the orientation.

If the distance of the wall is max_distance and scale == max_distance the returned value is 1. The stepsize (in pixels) controls the granularity at which the distance to the wall is computed.

FlyRL.FourWallsEncoder — Method

FourWallsEncoder(; max_distance = 40, scale = max_distance, stepsize = 1.5)

Encodes (x, y, ox, oy)-position-orientation tuples to relative distances to the next wall at 0, 90, 180 and 270 degrees relative to the orientation.

If the distance of the wall is max_distance and scale == max_distance the returned value is 1. The stepsize (in pixels) controls the granularity at which the distance to the wall is computed.

FlyRL.SixWallsEncoder — Method

SixWallsEncoder(; max_distance = 40, scale = max_distance, stepsize = 1.5)

Encodes (x, y, ox, oy)-position-orientation tuples to relative distances to the next wall at 0, 45, 135, 180, 225 and 315 degrees relative to the orientation.

If the distance of the wall is max_distance and scale == max_distance the returned value is 1. The stepsize (in pixels) controls the granularity at which the distance to the wall is computed.

FlyRL.decode — Method

decode(e, vectors)

Inverse of encode(VectorEncoder(), track).

FlyRL.encode! — Method

Encode track with encoder and store result in data frame track.

FlyRL.encode — Method

Encode track with encoder.