FlyRL.ChangeOf — Type

ChangeOf(stat)
ChangeOf(stat, midpoint)

Compute the change in statistics stat before and after the midpoint (= 0.5 by default).

Example

summarize(ChangeOf(RelativeTimeInShockArm()), track)

FlyRL.RelativeTimeInState — Method

RelativeTimeInState(encoder, state; exclude)

Compute the relative duration spent in state in the encoding defined by the encoder. States can be excluded with exclude = ["state1", "state2"].

Example

summarize(RelativeTimeInState(SemanticEncoder7(), "left turn", exclude = ["center"]), track)

FlyRL.RelativeVisitsToState — Method

RelativeVisitsToState(e, key; exclude)

Compute the relative number of visits to state in the encoding defined by the encoder. States can be excluded with exclude = ["state1", "state2"].

Example

summarize(RelativeVisitsToState(SemanticEncoder7(), "left turn", exclude = ["center"]), track)

FlyRL.RelativeTimeInShockArm — Method

RelativeTimeInShockArm()

Relative time in shock arm.

FlyRL.RelativeVisitsToShockArm — Method

RelativeVisitsToShockArm()

Relative visits to shock arm.